// src/components/Forms/PicPayGroup/tipo-demanda/interfaces.ts
export interface DemandaOutroPicPay {
  pk?: string;
  usuario_atuando?: string;
  created_by?: string;
  data_carimbo?: string;
  due_date?: string;
  arquivos?: ArquivoDemanda[];
  tipo_demanda?: string;
  status?: string;
  status_demanda?: string;
  perfil_demanda?: string;
  processo?: string;
  identificacao?: string;
  cliente?: string;
  observacao?: {
    mensagem: string;
    criada_em: string;
    criada_por: string;
  }[];
}

export interface FormFile {
  id: string;
  info: {
    name: string;
    size: number;
    sizeFormatted: string;
    url: string; 
    unit: Unit;
    s3Key: string;
    s3Bucket: string;
  };
  status: {
    success: boolean | undefined;
    progress: number;
    message: string | undefined;
    isDeleting?: boolean;
  };
  allow: {
    delete: boolean;
    download: boolean;
    retryUpload: boolean;
    link: boolean;
  };
  dropzoneFile: File | FileRejection | undefined;
}